refactor logic into separate class
https://coveord.atlassian.net/browse/JSUI-2594
